backend/functions/documents/requirements.txt
functions-framework==3.*
flask==2.3.3
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.29.0
firebase-admin==6.2.0
anthropic==0.17.0
pdf2image==1.16.3
poppler-utils==0.1.0
pillow==10.2.0
google-auth==2.40.1
werkzeug==2.3.7
